+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to the "hex-casting" extension will be documented in this file.
 
-## [unreleased]
+## [0.0.11] - 2022-12-19
 
 ### Additions
 
 - Add macro support with `#define`
 - Add diagnostics to warn on unknown patterns or invalid `#` directives
+- Add inlay hints for macros
 
 ### Fixes